requirements.txt
torchvision~=0.8.1
matplotlib~=3.3.2
numpy~=1.19.2
setuptools~=50.3.0
forknet~=0.1
click~=7.1.2
tqdm~=4.52.0
nibabel~=3.2.0
kornia~=0.3.0
ray-tune~=1.0.1
scikit-learn~=0.23.2
pandas~=1.1.3